:bricks: Refactor causes section
Move causes section to own section
shared components are not refactor, like Lightbox or modal Component because are used in other components
