+++ src/App.js
@@ -18,14 +18,13 @@ import WaitingPortBanner from './assets/images/Banners/Mobile/waiting-banner.png
 import WaitingPortBanner2 from './assets/images/Banners/Mobile/waiting.png'
 import WaitingTime from './pages/WaitingTime/WaitingTime'
 import Meanwhile from './pages/WaitingTime/Meanwhile'
-import CausesBanner from './assets/images/Causes/causes-banner.png'
+import CausesBanner from 'causes/assets/images/gallery/causes-banner.png'
 import CausesPortBanner from './assets/images/Banners/Mobile/causes-banner.png'
 import CausesPortBanner2 from './assets/images/Banners/Mobile/causes.png'
-import CausesBannerText from './assets/images/Causes/green_banner_text.png'
 import CausesDrawingsBanner from './assets/images/Banners/banner-drawings.png'
 import CausesStoryBanner from './assets/images/Banners/banner-story.png'
 import CausesSect from './causes/pages/Section/index'
-import CausesGallery from './pages/Gallery/index'
+import CausesGallery from './causes/pages/Gallery/index'
 import DreamsBanner from './assets/images/Dreams/dreams-banner.png'
 import PodcastBanner from './assets/images/Banners/dream-podcast.png'
 import DreamsGalleryBanner from './assets/images/Banners/dreams-gallery.png'
@@ -34,7 +33,6 @@ import DreamsPortBanner2 from './assets/images/Banners/Mobile/dreams.png'
 import DreamsGallery from './pages/DreamsGallery'
 import Dreams from './pages/Dreams/index'
 import WaitingTimeGallery from './pages/WaitingTime/WaitingTimeGallery'
-import WaitingSmallBanner from './assets/images/Banners/wait-s-banner.png'
 import WaitingGalleryBanner from './assets/images/Banners/wait-gallery-banner.png'
 import WaitingProcessBanner from './assets/images/Banners/wait-process-banner.png'
 import WaitingVideosBanner from './assets/images/Banners/wait-videos-banner.png'
@@ -44,7 +42,6 @@ import Poadcast from './pages/Dreams/Poadcast'
 import { Helmet } from 'react-helmet'
 import './assets/css/index.css'
 import Comic from './causes/pages/Comic/FlipPage'
-import NotesImages from './pages/Support/Concepts/NotesImages'
 import NotFound from './pages/NotFound'
 
 function App() {
