+++ src/causes/assets/images/comic/cover.png
