+++ src/causes/assets/images/comic/layer_2.png
