+++ src/causes/assets/images/comic/layer_6.png
