+++ src/causes/assets/images/comic/layer_7.png
