+++ src/causes/assets/styles/index.css
@@ -3,3 +3,4 @@
   filter: grayscale(100%);
   transition: filter 0.2s ease-out;
 }
+
