+++ src/causes/components/Comic/Comic.js
