+++ src/causes/components/Comic/InnerButtons.js
