+++ src/causes/components/Gallery/Bullet/Bullet.js
